jupiter_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
python-jose
stripe
pydantic-settings
greenlet
passlib
python-multipart
fastapi-mail
bcrypt